Update voidwatch.lua
Added some minor changes for multiboxing.  User setting files will now be created and a new command "altdelay" was added so that a different range of delays can be added for each character (currently only applies to opening the chest).  
The altdelay range is 0-5, just give each character a unique number from 0-5 and you should be good to go.